backend/requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgspec==0.19.0
numpy==2.2.2
oauthlib==3.2.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
